update UserAdminPitch and fix bug
