+++ src/layouts/AdminLayout.tsx
@@ -283,7 +283,7 @@ const AdminLayout = () => {
                   className="border-b-0 p-3"
                 >
                   <Link
-                    to={"/admin/location"}
+                    to={"/admin/comment"}
                     className="flex items-center gap-2"
                   >
                     <ListItemPrefix>
@@ -306,7 +306,7 @@ const AdminLayout = () => {
                       color="blue-gray"
                       className="mr-auto font-normal"
                     >
-                      Địa Chỉ
+                      Bình Luận
                     </Typography>
                   </Link>
                 </AccordionHeader>
